Refactored to remove superficial package layers; eliminated duplicate code; misc fixes
