+++ pkg/store/store.go
@@ -1,4 +1,4 @@
-// Copyright 2019-present Open Networking Foundation.
+// Copyright 2021-present Open Networking Foundation.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package topo
+package store
 
 import (
 	"context"
@@ -232,7 +232,7 @@ func (s *atomixStore) Watch(ctx context.Context, ch chan<- topoapi.Event, filter
 	}
 
 	mapCh := make(chan *_map.Event)
-	if err := s.objects.Watch(context.Background(), mapCh, watchOpts...); err != nil {
+	if err := s.objects.Watch(ctx, mapCh, watchOpts...); err != nil {
 		return errors.FromAtomix(err)
 	}
